# backend/docs/swagger.yaml
basePath: /
definitions:
...
  backend_internal_types.CreateBrandRequest:
    properties:
      brand_name:
        type: string
    type: object
  backend_internal_types.CreateBrandResponse:
    properties:
      brand_id:
        type: integer
    type: object
...
  backend_internal_types.DeleteBrandRequest:
    properties:
      brand_id:
        type: integer
    type: object
  backend_internal_types.DeleteBrandResponse:
    properties:
      status:
        type: string
    type: object
...
  backend_internal_types.GetBrandByIdRequest:
    properties:
      brand_id:
        type: integer
    type: object
  backend_internal_types.GetBrandByIdResponse:
    properties:
      brand_name:
        type: string
    type: object
...
  backend_internal_types.GetIdByBrandRequest:
    properties:
      brand_name:
        type: string
    type: object
  backend_internal_types.GetIdByBrandResponse:
    properties:
      brand_id:
        type: integer
    type: object
...
  backend_internal_types.UpdateBrandRequest:
    properties:
      new_brand_name:
        type: string
      old_brand_name:
        type: string
    type: object
  backend_internal_types.UpdateBrandResponse:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: API server for stream service application
  title: Stream Service Prototype API
  version: "1.0"
paths:
...
  /api/brand/create-brand:
    post:
      consumes:
      - application/json
      description: creating brand
      operationId: create-brand
      parameters:
      - description: Создаёт бренд
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_internal_types.CreateBrandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_types.CreateBrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_types.CreateBrandResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_types.CreateBrandResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_types.CreateBrandResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/backend_internal_types.CreateBrandResponse'
      summary: createBrand
      tags:
      - Brand
  /api/brand/delete-brand:
    delete:
      consumes:
      - application/json
      description: deleting brand
      operationId: delete-brand
      parameters:
      - description: Удаляет бренд
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_internal_types.DeleteBrandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_types.DeleteBrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_types.DeleteBrandResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_types.DeleteBrandResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_types.DeleteBrandResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/backend_internal_types.DeleteBrandResponse'
      summary: deleteBrand
      tags:
      - Brand
  /api/brand/get-brand-by-id:
    get:
      consumes:
      - application/json
      description: getting brand id by brand name
      operationId: get-brand-name-by-brand-id
      parameters:
      - description: Возвращает название бренда по id
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_internal_types.GetBrandByIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_types.GetBrandByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_types.GetBrandByIdResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_types.GetBrandByIdResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_types.GetBrandByIdResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/backend_internal_types.GetBrandByIdResponse'
      summary: getBrandById
      tags:
      - Brand
  /api/brand/get-id-by-brand:
    get:
      consumes:
      - application/json
      description: getting brand id by brand name
      operationId: get-brand-id-by-brand-name
      parameters:
      - description: Возвращает id бренда по названию
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_internal_types.GetIdByBrandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_types.GetIdByBrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_types.GetIdByBrandResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_types.GetIdByBrandResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_types.GetIdByBrandResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/backend_internal_types.GetIdByBrandResponse'
      summary: getIdByBrand
      tags:
      - Brand
  /api/brand/update-brand:
    put:
      consumes:
      - application/json
      description: updating brand
      operationId: update-brand
      parameters:
      - description: Обновляет бренд
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_internal_types.UpdateBrandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_types.UpdateBrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_types.UpdateBrandResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_types.UpdateBrandResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_types.UpdateBrandResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/backend_internal_types.UpdateBrandResponse'
      summary: updateBrand
      tags:
      - Brand
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"